Move aggregator logic to /generate route handling
